+++ gyrh-go-v2/frontend/src/pages/kiosk/KioskViewer.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { AppShell } from '../../app/AppShell';
+import { kioskScreens } from '../../constants';
+
+export function KioskViewer() {
+  return <AppShell mode="kiosk" navigationItems={kioskScreens} />;
+}